1.byom-imgclassification-from-scratch/artifacts/requirements.txt
wheel
Cython
numpy==1.19.4
dlr==1.8.0
awsiotsdk
pillow